Display source as GFW on gear type and vessel type
