+++ applications/vessel-history/src/features/profile/components/InfoField.tsx
@@ -4,32 +4,10 @@ import { useTranslation } from 'react-i18next'
 import I18nDate from 'features/i18n/i18nDate'
 import { DEFAULT_EMPTY_VALUE } from 'data/config'
 import { ValueItem } from 'types'
-import styles from './Info.module.css'
+import { VesselFieldLabel } from 'types/vessel'
 import InfoFieldHistory from './InfoFieldHistory'
+import styles from './Info.module.css'
 
-export enum VesselFieldLabel {
-  name = 'name',
-  flag = 'flag',
-  shipname = 'shipname',
-  firstTransmissionDate = 'firstTransmissionDate',
-  lastTransmissionDate = 'lastTransmissionDate',
-  imo = 'imo',
-  mmsi = 'mmsi',
-  callsign = 'callsign',
-  fleet = 'fleet',
-  origin = 'origin',
-  type = 'type',
-  geartype = 'geartype',
-  length = 'length',
-  depth = 'depth',
-  grossTonnage = 'grossTonnage',
-  owner = 'owner',
-  operator = 'operator',
-  builtYear = 'builtYear',
-  authorizations = 'authorizations',
-  registeredGearType = 'registeredGearType',
-  iuuStatus = 'iuuStatus',
-}
 interface ListItemProps {
   label: VesselFieldLabel
   value?: string
