+++ applications/vessel-history/src/features/vessels/vessels.hook.ts
@@ -2,17 +2,22 @@ import { useCallback } from 'react'
 import { useTranslation } from 'react-i18next'
 import { DEFAULT_EMPTY_VALUE } from 'data/config'
 import { VesselAPISource } from 'types'
+import { VesselFieldLabel } from 'types/vessel'
 
-export const useVesselsConnect = () => {
+export const useVesselsConnect = (field?: VesselFieldLabel) => {
   const { t } = useTranslation()
+  const gfwSourceLabel =
+    field && [VesselFieldLabel.geartype, VesselFieldLabel.type].includes(field)
+      ? t('common.GFW', 'GFW')
+      : t('common.AIS', 'AIS')
   const formatSource = useCallback(
     (source?: VesselAPISource) =>
       source
         ? source === VesselAPISource.GFW
-          ? t('common.AIS', 'AIS')
+          ? gfwSourceLabel
           : t('common.other', 'Other')
         : DEFAULT_EMPTY_VALUE,
-    [t]
+    [gfwSourceLabel, t]
   )
   return {
     formatSource,
